+++ spec/unit/show-hide-content.spec.js
@@ -32,6 +32,18 @@ describe('show-hide-content', function () {
         '<div id="show-hide-radios" class="panel js-hidden" />' +
         '</form>' +
 
+        '<form>' +
+        '<div class="multiple-choice" data-target="show-hide-radios--name-with-characters-to-escape">' +
+        '<input type="radio" name="test.single[option1=\'value1\'][option2=\'value2\']" value="yes">' +
+        '<label>Yes</label>' +
+        '</div>' +
+        '<div class="multiple-choice">' +
+        '<input type="radio" name="test.single[option1=\'value1\'][option2=\'value2\']" value="no">' +
+        '<label>No</label>' +
+        '</div>' +
+        '<div id="show-hide-radios--name-with-characters-to-escape" class="panel js-hidden" />' +
+        '</form>' +
+
         // Checkboxes (multiple values)
         '<form>' +
         '<div class="multiple-choice" data-target="show-hide-checkboxes">' +
@@ -58,6 +70,10 @@ describe('show-hide-content', function () {
       this.$radio1 = $radios.eq(0)
       this.$radio2 = $radios.eq(1)
 
+      // Two more radios, with name attributes that include periods
+      this.$radio3 = $radios.eq(2)
+      this.$radio4 = $radios.eq(3)
+
       // Three checkboxes
       this.$checkbox1 = $checkboxes.eq(0)
       this.$checkbox2 = $checkboxes.eq(1)
@@ -68,6 +84,7 @@ describe('show-hide-content', function () {
 
       // Show/Hide content
       this.$radioShowHide = $('#show-hide-radios')
+      this.$radioShowHide_nameWithCharactersToEscape = $('#show-hide-radios--name-with-characters-to-escape')
       this.$checkboxShowHide = $('#show-hide-checkboxes')
 
       // Add show/hide content support
@@ -79,25 +96,36 @@ describe('show-hide-content', function () {
       it('should add the aria attributes to inputs with show/hide content', function () {
         expect(this.$radio1.attr('aria-expanded')).toBe('false')
         expect(this.$radio1.attr('aria-controls')).toBe('show-hide-radios')
+
+        expect(this.$radio3.attr('aria-expanded')).toBe('false')
+        expect(this.$radio3.attr('aria-controls')).toBe('show-hide-radios--name-with-characters-to-escape')
       })
 
       it('should add the aria attributes to show/hide content', function () {
         expect(this.$radioShowHide.attr('aria-hidden')).toBe('true')
         expect(this.$radioShowHide.hasClass('js-hidden')).toEqual(true)
+
+        expect(this.$radioShowHide_nameWithCharactersToEscape.attr('aria-hidden')).toBe('true')
+        expect(this.$radioShowHide_nameWithCharactersToEscape.hasClass('js-hidden')).toEqual(true)
       })
 
       it('should hide the show/hide content visually', function () {
         expect(this.$radioShowHide.hasClass('js-hidden')).toEqual(true)
+        expect(this.$radioShowHide_nameWithCharactersToEscape.hasClass('js-hidden')).toEqual(true)
       })
 
       it('should do nothing if no radios are checked', function () {
         expect(this.$radio1.attr('aria-expanded')).toBe('false')
         expect(this.$radio2.attr('aria-expanded')).toBe(undefined)
+        expect(this.$radio3.attr('aria-expanded')).toBe('false')
+        expect(this.$radio4.attr('aria-expanded')).toBe(undefined)
       })
 
       it('should do nothing if no checkboxes are checked', function () {
         expect(this.$radio1.attr('aria-expanded')).toBe('false')
         expect(this.$radio2.attr('aria-expanded')).toBe(undefined)
+        expect(this.$radio3.attr('aria-expanded')).toBe('false')
+        expect(this.$radio4.attr('aria-expanded')).toBe(undefined)
       })
 
       describe('with non-default markup', function () {
@@ -107,12 +135,18 @@ describe('show-hide-content', function () {
 
         it('should do nothing if a radio without show/hide content is checked', function () {
           this.$radio2.prop('checked', true)
+          this.$radio4.prop('checked', true)
 
           // Defaults changed, initialise again
           this.showHideContent = new GOVUK.ShowHideContent().init()
           expect(this.$radio1.attr('aria-expanded')).toBe('false')
+          expect(this.$radio3.attr('aria-expanded')).toBe('false')
+
           expect(this.$radioShowHide.attr('aria-hidden')).toBe('true')
+          expect(this.$radioShowHide_nameWithCharactersToEscape.attr('aria-hidden')).toBe('true')
+
           expect(this.$radioShowHide.hasClass('js-hidden')).toEqual(true)
+          expect(this.$radioShowHide_nameWithCharactersToEscape.hasClass('js-hidden')).toEqual(true)
         })
 
         it('should do nothing if a checkbox without show/hide content is checked', function () {
@@ -138,12 +172,18 @@ describe('show-hide-content', function () {
 
         it('should make the show/hide content visible if its radio is checked', function () {
           this.$radio1.prop('checked', true)
+          this.$radio3.prop('checked', true)
 
           // Defaults changed, initialise again
           this.showHideContent = new GOVUK.ShowHideContent().init()
           expect(this.$radio1.attr('aria-expanded')).toBe('true')
+          expect(this.$radio3.attr('aria-expanded')).toBe('true')
+
           expect(this.$radioShowHide.attr('aria-hidden')).toBe('false')
+          expect(this.$radioShowHide_nameWithCharactersToEscape.attr('aria-hidden')).toBe('false')
+
           expect(this.$radioShowHide.hasClass('js-hidden')).toEqual(false)
+          expect(this.$radioShowHide_nameWithCharactersToEscape.hasClass('js-hidden')).toEqual(false)
         })
 
         it('should make the show/hide content visible if its checkbox is checked', function () {
@@ -160,14 +200,23 @@ describe('show-hide-content', function () {
       describe('and a show/hide radio receives a click', function () {
         it('should make the show/hide content visible', function () {
           this.$radio1.click()
+          this.$radio3.click()
           expect(this.$radioShowHide.hasClass('js-hidden')).toEqual(false)
+          expect(this.$radioShowHide_nameWithCharactersToEscape.hasClass('js-hidden')).toEqual(false)
         })
 
         it('should add the aria attributes to show/hide content', function () {
           this.$radio1.click()
+          this.$radio3.click()
+
           expect(this.$radio1.attr('aria-expanded')).toBe('true')
+          expect(this.$radio3.attr('aria-expanded')).toBe('true')
+
           expect(this.$radioShowHide.attr('aria-hidden')).toBe('false')
+          expect(this.$radioShowHide_nameWithCharactersToEscape.attr('aria-hidden')).toBe('false')
+
           expect(this.$radioShowHide.hasClass('js-hidden')).toEqual(false)
+          expect(this.$radioShowHide_nameWithCharactersToEscape.hasClass('js-hidden')).toEqual(false)
         })
       })
 
@@ -225,6 +274,10 @@ describe('show-hide-content', function () {
         }))
         expect(events && events.click).toContain(jasmine.objectContaining({
           namespace: 'ShowHideContent',
+          selector: 'input[type="radio"][name="test.single[option1=\'value1\'][option2=\'value2\']"]'
+        }))
+        expect(events && events.click).toContain(jasmine.objectContaining({
+          namespace: 'ShowHideContent',
           selector: '[data-target] > input[type="checkbox"]'
         }))
       })
@@ -240,6 +293,10 @@ describe('show-hide-content', function () {
         expect(events && events.click).not.toContain(jasmine.objectContaining({
           namespace: 'ShowHideContent',
           selector: 'input[type="radio"][name="single"]'
+        }))
+        expect(events && events.click).not.toContain(jasmine.objectContaining({
+          namespace: 'ShowHideContent',
+          selector: 'input[type="radio"][name="test.single[option1=\'value1\'][option2=\'value2\']"]'
         }))
         expect(events && events.click).not.toContain(jasmine.objectContaining({
           namespace: 'ShowHideContent',